+++ config/routes.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
-GemTemplate::Engine.routes.draw do
+RecordingStudioPublishedArticle::Engine.routes.draw do
+  get "recordings/:recording_id/articles", to: "articles#index", as: :recording_articles
+  get "articles/:id", to: "articles#show", as: :article
   root "home#index"
 end