+++ test/dummy/config/initializers/recording_studio_publications.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+RecordingStudioPublications.configure do |config|
+end